add Linkedlist and addTask
Linkedlist: I chose this datastructure because I'm dealing with inserting and deleting and modifying elements.
listing them would be also more efficient with this.

addTask: this function will be used in TaskCLI to add a task with command line
initialises an object class and puts it in linkedlist (tasks)
gives an id and description